model_package/requirements/requirements.txt
torch==2.0.1
numpy==1.25.0
tqdm==4.65.0
transformers==4.33.2
lightning==2.3.1
pillow==9.5.0
strictyaml==1.7.3
ruamel.yaml==0.18.6
pydantic==1.8.2
